# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HR & Payroll Hub",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/market": {
      "get": {
        "tags": [
          "market"
        ],
        "summary": "Get Market Data",
        "operationId": "get_market_data_api_market_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MarketDataResponse"
                  },
                  "type": "array",
                  "title": "Response Get Market Data Api Market Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "market"
        ],
        "summary": "Create Market Data",
        "operationId": "create_market_data_api_market_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarketDataCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketDataResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/market/{id}/entries": {
      "get": {
        "tags": [
          "market"
        ],
        "summary": "Get Market Entries",
        "operationId": "get_market_entries_api_market__id__entries_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MarketEntryResponse"
                  },
                  "title": "Response Get Market Entries Api Market  Id  Entries Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/market/entries": {
      "post": {
        "tags": [
          "market"
        ],
        "summary": "Add Market Entry",
        "operationId": "add_market_entry_api_market_entries_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarketEntryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketEntryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/market/entries/{id}": {
      "delete": {
        "tags": [
          "market"
        ],
        "summary": "Delete Market Entry",
        "operationId": "delete_market_entry_api_market_entries__id__delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/market/{id}": {
      "delete": {
        "tags": [
          "market"
        ],
        "summary": "Delete Market Data",
        "operationId": "delete_market_data_api_market__id__delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/market/{id}/sync-hh": {
      "post": {
        "tags": [
          "market"
        ],
        "summary": "Sync Market With Hh",
        "operationId": "sync_market_with_hh_api_market__id__sync_hh_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MarketDataCreate": {
        "properties": {
          "position_title": {
            "type": "string",
            "title": "Position Title"
          },
          "branch_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch Id"
          },
          "min_salary": {
            "type": "integer",
            "title": "Min Salary",
            "default": 0
          },
          "max_salary": {
            "type": "integer",
            "title": "Max Salary",
            "default": 0
          },
          "median_salary": {
            "type": "integer",
            "title": "Median Salary",
            "default": 0
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "position_title"
        ],
        "title": "MarketDataCreate"
      },
      "MarketDataResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "position_title": {
            "type": "string",
            "title": "Position Title"
          },
          "branch_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch Id"
          },
          "min_salary": {
            "type": "integer",
            "title": "Min Salary"
          },
          "max_salary": {
            "type": "integer",
            "title": "Max Salary"
          },
          "median_salary": {
            "type": "integer",
            "title": "Median Salary"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/MarketPointResponse"
            },
            "type": "array",
            "title": "Points"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "position_title",
          "min_salary",
          "max_salary",
          "median_salary",
          "updated_at"
        ],
        "title": "MarketDataResponse"
      },
      "MarketEntryCreate": {
        "properties": {
          "market_id": {
            "type": "integer",
            "title": "Market Id"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "salary": {
            "type": "integer",
            "title": "Salary"
          }
        },
        "type": "object",
        "required": [
          "market_id",
          "company_name",
          "salary"
        ],
        "title": "MarketEntryCreate"
      },
      "MarketEntryResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "market_id": {
            "type": "integer",
            "title": "Market Id"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "salary": {
            "type": "integer",
            "title": "Salary"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "market_id",
          "company_name",
          "salary",
          "created_at"
        ],
        "title": "MarketEntryResponse"
      },
      "MarketPointResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "market_id": {
            "type": "integer",
            "title": "Market Id"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "salary": {
            "type": "integer",
            "title": "Salary"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "market_id",
          "company_name",
          "salary",
          "created_at"
        ],
        "title": "MarketPointResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}